requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Faker==13.14.0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
outcome==1.2.0
packaging==22.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.1.0
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
urllib3-secure-extra==0.1.0
webdriver-manager==3.7.1
wsproto==1.2.0
